requirements.txt
# Core Vanna (no Gemini now)
vanna[sqlite]>=2.0.0

# Groq LLM
groq

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data / charts
plotly>=5.20.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0